Reduce visibility: no part of the public API (yet),
