+++ src/CodeAnalysis.TestTools/Contexts/AnalyzerVerifyContext.cs
@@ -30,7 +30,7 @@ public abstract record AnalyzerVerifyContext
     /// <summary>Gets the diagnostics.</summary>
     [Pure]
     public async Task<IReadOnlyCollection<Diagnostic>> GetDiagnosticsAsync()
-        => await (await GetCompilationAsync()).GetDiagnosticsAsync(Analyzers, GetTextDocuments());
+        => await (await GetCompilationAsync()).GetDiagnosticsAsync(Analyzers, GetAdditionalText());
 
     /// <summary>Reports (both expected, unexpected, and not reported) issues for the analyzer verify context.</summary>
     [Pure]
@@ -56,5 +56,5 @@ public abstract record AnalyzerVerifyContext
 
     /// <summary>Gets the additional texts.</summary>
     [Pure]
-    protected abstract IEnumerable<TextDocument> GetTextDocuments();
+    internal abstract IEnumerable<AdditionalText> GetAdditionalText();
 }
